+++ elm-format-lib/src/Box.hs
@@ -2,15 +2,19 @@
 module Box
   ( Line, identifier, keyword, punc, literal, row, space
   , Box(SingleLine, MustBreak), blankLine, line, mustBreak, stack', stack1, andThen
-  , isLine, allSingles, lineLength
+  , isLine, allSingles
   , indent, prefix, addSuffix
   , render
-  ) where
+  ,allSingles2,allSingles3,lineLength,isSingle,isMustBreak,comment,stack,joinMustBreak,prefixOrIndent) where
 
 import Data.Fix
-import Elm.Utils ((|>), List)
+import Elm.Utils (List)
 
 import qualified Data.Text as T
+import Indent (Indent)
+import qualified Indent
+import Data.Semigroup (sconcat)
+import Data.List.NonEmpty (NonEmpty((:|)))
 
 
 {-
@@ -23,42 +27,56 @@ Space is self-explanatory,
 -}
 data LineF a
     = Text T.Text
-    | Row (List a)
+    | Row a a
     | Space
-    | Tab
 
 type Line = Fix LineF
 
+instance Semigroup Line where
+    a <> b = Fix $ Row a b
 
-identifier :: String -> Line
+
+identifier :: T.Text -> Line
 identifier =
-    Fix . Text . T.pack
+    Fix . Text
 
 
-keyword :: String -> Line
+keyword :: T.Text -> Line
 keyword =
-    Fix . Text . T.pack
+    Fix . Text
 
 
-punc :: String -> Line
+punc :: T.Text -> Line
 punc =
-    Fix . Text . T.pack
+    Fix . Text
 
 
-literal :: String -> Line
+literal :: T.Text -> Line
 literal =
-    Fix . Text . T.pack
+    Fix . Text
 
 
--- join more Line elements into one
+comment :: T.Text -> Line
+comment =
+    Fix . Text
+
+
+{-# DEPRECATED row "use `(<>)` instead" #-}
+{-| join more Line elements into one
+-}
 row :: List Line -> Line
-row =
-    Fix . Row
+row [] = error $ "<elm-format: " ++ "UNEXPECTED ROW" ++ ": " ++ "no elements" ++ " -- please report this at https://github.com/avh4/elm-format/issues >"
+row (first:rest) = sconcat (first:|rest)
 
 
 space :: Line
 space =
     Fix Space
+
+
+data Indented a =
+    Indented Indent a
+    deriving (Functor)
 
 
 {-
@@ -77,9 +95,9 @@ Sometimes (see `prefix`) the first line of Stack
   gets different treatment than the other lines.
 -}
 data Box
-    = SingleLine Line
-    | Stack Line Line [Line]
-    | MustBreak Line
+    = SingleLine (Indented Line)
+    | Stack (Indented Line) (Indented Line) [Indented Line]
+    | MustBreak (Indented Line)
 
 
 blankLine :: Box
@@ -89,12 +107,21 @@ blankLine =
 
 line :: Line -> Box
 line =
-    SingleLine
+    SingleLine . mkIndentedLine
 
 
 mustBreak :: Line -> Box
 mustBreak =
-    MustBreak
+    MustBreak . mkIndentedLine
+
+
+mkIndentedLine :: Line -> Indented Line
+mkIndentedLine (Fix Space) = Indented (Indent.spaces 1) (literal "")
+mkIndentedLine (Fix (Row (Fix Space) next)) =
+    let (Indented i rest') = mkIndentedLine next
+    in
+    Indented (Indent.spaces 1 <> i) rest'
+mkIndentedLine other = Indented mempty other
 
 
 stack' :: Box -> Box -> Box
@@ -114,6 +141,11 @@ andThen rest first =
     foldl stack' first rest
 
 
+stack :: Box -> [Box] -> Box
+stack first rest = stack1 (first:rest)
+
+
+{-# DEPRECATED stack1 "Prefer `stack` or `stack'`" #-}
 stack1 :: [Box] -> Box
 stack1 children =
     case children of
@@ -125,12 +157,42 @@ stack1 children =
             foldr1 stack' boxes
 
 
-mapLines :: (Line -> Line) -> Box -> Box
+joinMustBreak :: Box -> Box -> Box
+joinMustBreak inner eol =
+    case (inner, eol) of
+        (SingleLine (Indented i1 inner'), SingleLine (Indented _ eol')) ->
+            SingleLine $ Indented i1 $
+            inner' <> space <> eol'
+
+        (SingleLine (Indented i1 inner'), MustBreak (Indented _ eol')) ->
+            MustBreak $ Indented i1 $
+            inner' <> space <> eol'
+
+        _ ->
+            stack' inner eol
+
+
+prefixOrIndent :: Box -> Box -> Box
+prefixOrIndent a b =
+    case ( a, b ) of
+        (SingleLine (Indented i1 a'), SingleLine (Indented _ b')) ->
+            SingleLine $ Indented i1 $
+            a' <> space <> b'
+
+        (SingleLine (Indented i1 a'), MustBreak (Indented _ b')) ->
+            MustBreak $ Indented i1 $
+            a' <> space <> b'
+
+        _ ->
+            stack' a (indent b)
+
+
+mapLines :: (Indented Line -> Indented Line) -> Box -> Box
 mapLines fn =
     mapFirstLine fn fn
 
 
-mapFirstLine :: (Line -> Line) -> (Line -> Line) -> Box -> Box
+mapFirstLine :: (Indented Line -> Indented Line) -> (Indented Line -> Indented Line) -> Box -> Box
 mapFirstLine firstFn restFn b =
     case b of
         SingleLine l1 ->
@@ -141,21 +203,51 @@ mapFirstLine firstFn restFn b =
             MustBreak (firstFn l1)
 
 
+mapLastLine :: (Indented Line -> Indented Line) -> Box -> Box
+mapLastLine lastFn = \case
+    SingleLine l1 ->
+        SingleLine (lastFn l1)
+    Stack l1 l2 [] ->
+        Stack l1 (lastFn l2) []
+    Stack l1 l2 ls ->
+        Stack l1 l2 (init ls ++ [lastFn $ last ls])
+    MustBreak l1 ->
+        MustBreak (lastFn l1)
+
+
 indent :: Box -> Box
 indent =
-    mapLines (\l -> row [Fix Tab, l])
+    mapLines (\(Indented i l) -> Indented (Indent.tab <> i) l)
 
 
 isLine :: Box -> Either Box Line
 isLine b =
     case b of
-        SingleLine l ->
+        SingleLine (Indented _ l) ->
             Right l
         _ ->
             Left b
 
 
-destructure :: Box -> (Line, [Line])
+isSingle :: Box -> Maybe Line
+isSingle b =
+    case b of
+        SingleLine (Indented _ l) ->
+            Just l
+        _ ->
+            Nothing
+
+
+isMustBreak :: Box -> Maybe Line
+isMustBreak b =
+    case b of
+        MustBreak (Indented _ l) ->
+            Just l
+        _ ->
+            Nothing
+
+
+destructure :: Box -> (Indented Line, [Indented Line])
 destructure b =
     case b of
         SingleLine l1 ->
@@ -166,7 +258,7 @@ destructure b =
             (l1, [])
 
 
-allSingles :: [Box] -> Either [Box] [Line]
+allSingles :: Traversable t => t Box -> Either (t Box) (t Line)
 allSingles boxes =
     case mapM isLine boxes of
         Right lines' ->
@@ -175,9 +267,25 @@ allSingles boxes =
             Left boxes
 
 
+allSingles2 :: Box -> Box -> Either (Box, Box) (Line, Line)
+allSingles2 b1 b2 =
+    case allSingles [b1, b2] of
+        Right [l1, l2] -> Right (l1, l2)
+        _ -> Left (b1, b2)
+
+
+allSingles3 :: Box -> Box -> Box -> Either (Box, Box, Box) (Line, Line, Line)
+allSingles3 b1 b2 b3 =
+    case allSingles [b1, b2, b3] of
+        Right [l1, l2, l3] -> Right (l1, l2, l3)
+        _ -> Left (b1, b2, b3)
+
+
 {-
 Add the prefix to the first line,
 pad the other lines with spaces of the same length
+
+NOTE: An exceptional case that we haven't really designed for is if the first line of the input Box is indented.
 
 EXAMPLE:
 abcde
@@ -189,127 +297,44 @@ myPrefix abcde
 prefix :: Line -> Box -> Box
 prefix pref =
     let
-        prefixLength = lineLength 0 pref
-        paddingSpaces = replicate prefixLength space
-        padLineWithSpaces l = row [ row paddingSpaces, l ]
-        addPrefixToLine l = row [ pref, l ]
+        prefixLength = fromIntegral $ T.length $ renderLine pref
+        padLineWithSpaces (Indented i l) = Indented (Indent.spaces prefixLength <> i) l
+        addPrefixToLine l = pref <> l
     in
-      mapFirstLine addPrefixToLine padLineWithSpaces
+    mapFirstLine (fmap addPrefixToLine) padLineWithSpaces
 
 
 addSuffix :: Line -> Box -> Box
-addSuffix suffix b =
-    case destructure b of
-        (l,[]) ->
-            line $ row [ l, suffix ]
-        (l1,ls) ->
-            line l1
-                |> andThen (map line $ init ls)
-                |> andThen [ line $ row [ last ls, suffix ] ]
+addSuffix suffix =
+    mapLastLine $ fmap (<> suffix)
 
 
-renderLine :: Int -> Line -> T.Text
-renderLine startColumn line' =
+renderIndentedLine :: Indented T.Text -> T.Text
+renderIndentedLine (Indented i line') =
+    T.replicate (Indent.width i) " " <> line'
+
+
+renderLine :: Line -> T.Text
+renderLine line' =
     case unFix line' of
         Text text ->
             text
         Space ->
             T.singleton ' '
-        Tab ->
-            T.pack $ replicate (tabLength startColumn) ' '
-        Row lines' ->
-            renderRow startColumn lines'
+        Row left right ->
+            renderLine left <> renderLine right
 
 
 render :: Box -> T.Text
 render box' =
     case box' of
         SingleLine line' ->
-            T.snoc (T.stripEnd $ renderLine 0 line') '\n'
+            T.snoc (T.stripEnd $ renderIndentedLine $ renderLine <$> line') '\n'
         Stack l1 l2 rest ->
-            T.unlines $ map (T.stripEnd . renderLine 0) (l1 : l2 : rest)
+            T.unlines $ map (T.stripEnd . renderIndentedLine . fmap renderLine) (l1 : l2 : rest)
         MustBreak line' ->
-            T.snoc (T.stripEnd $ renderLine 0 line') '\n'
+            T.snoc (T.stripEnd $ renderIndentedLine $ renderLine <$> line') '\n'
 
 
-lineLength :: Int -> Line -> Int
-lineLength startColumn line' =
-   startColumn +
-      case unFix line' of
-         Text string -> T.length string
-         Space -> 1
-         Tab -> tabLength startColumn
-         Row lines' -> rowLength startColumn lines'
-
-
-initRow :: Int -> (T.Text, Int)
-initRow startColumn =
-  (T.empty, startColumn)
-
-
-spacesInTab :: Int
-spacesInTab =
-  4
-
-
-spacesToNextTab :: Int -> Int
-spacesToNextTab startColumn =
-  startColumn `mod` spacesInTab
-
-tabLength :: Int -> Int
-tabLength startColumn =
-  spacesInTab - spacesToNextTab startColumn
-
-{-
-What happens here is we take a row and start building its contents
-  along with the resulting length of the string. We need to have that
-  because of Tabs, which need to be passed the current column in arguments
-  in order to determine how many Spaces are they going to span.
-  (See `tabLength`.)
-
-So for example if we have a Box [Space, Tab, Text "abc", Tab, Text "x"],
-  it goes like this:
-
-string      | column | todo
-""          | 0      | [Space, Tab, Text "abc", Tab, Text "x"]
-" "         | 1      | [Tab, Text "abc", Tab, Text "x"]
-"    "      | 4      | [Text "abc", Tab, Text "x"]
-"    abc"   | 7      | [Tab, Text "x"]
-"    abc "  | 8      | [Text "x"]
-"    abc x" | 9      | []
-
-Thus we get the result string with correctly rendered Tabs.
-
-The (T.Text, Int) type here means the (string, column) from the table above.
-
-Then we just need to do one final modification to get from endColumn to resultLength,
-  which is what we are after in the function `rowLength`.
--}
-renderRow' :: Int -> [Line] -> (T.Text, Int)
-renderRow' startColumn lines' =
-  (result, resultLength)
-  where
-    (result, endColumn) = foldl addLine (initRow startColumn) lines'
-    resultLength = endColumn - startColumn
-
-{-
-A step function for renderRow'.
-
-addLine (" ",1) Tab == ("    ",4)
--}
-addLine :: (T.Text, Int) -> Line -> (T.Text, Int)
-addLine (string, startColumn') line' =
-  (newString, newStartColumn)
-  where
-    newString = T.append string $ renderLine startColumn' line'
-    newStartColumn = lineLength startColumn' line'
-
--- Extract the final string from renderRow'
-renderRow :: Int -> [Line] -> T.Text
-renderRow startColumn lines' =
-  fst $ renderRow' startColumn lines'
-
--- Extract the final length from renderRow'
-rowLength :: Int -> [Line] -> Int
-rowLength startColumn lines' =
-  snd $ renderRow' startColumn lines'
+lineLength :: Line -> Int
+lineLength = T.length . renderLine