+++ elm-format-lib/src/ElmFormat/Render/Text.hs
@@ -11,11 +11,13 @@ import AST.Module (Module)
 import qualified Box
 import qualified Data.Text as Text
 import qualified ElmFormat.Render.Box as Render
+import qualified Data.Fix as Fix
+import qualified ElmFormat.Render.ElmStructure as ElmStructure
 
 
 render :: Coapplicative annf => ElmVersion -> Module [UppercaseIdentifier] (ASTNS annf [UppercaseIdentifier] 'TopLevelNK) -> Text.Text
 render elmVersion modu =
-    renderBox $ Render.formatModule elmVersion True 2 modu
+    renderBox $ Fix.cata ElmStructure.render $ Render.formatModule elmVersion True 2 modu
 
 
 renderBox :: Box.Box -> Text.Text
@@ -27,4 +29,4 @@ renderBox box =
 -- TODO: remove this and convert the Integration test to a test fixture
 renderLiteral :: ElmVersion -> LiteralValue -> Text.Text
 renderLiteral elmVersion literal =
-    renderBox $ Render.formatLiteral elmVersion literal
+    renderBox $ Fix.cata ElmStructure.render $ Render.formatLiteral elmVersion literal